enhance: return user limit and user count in version API (#7349)
We also return user limit violations in the licenseEntitlementViolations.

Since we now have the community license, this change also sets the
"enterprise" flag to true only when an enterprise license is actually
set.

This change also updates the error message displayed to a user when
they can't create an account because the user limit has been reached.

+++ pkg/gateway/client/auth.go
@@ -97,7 +97,7 @@ func (u UserDecorator) resolveUserLimit(ctx context.Context) (UserLimit, error) 
 // UserLimit describes the maximum number of users an installation may have.
 // Maximum is ignored when Unlimited is true.
 type UserLimit struct {
-	Maximum   int
+	Maximum   int64
 	Unlimited bool
 }
